Make analytics functions truly non-blocking with `tokio::spawn`
Make analytics functions truly non-blocking with tokio::spawn

Previously, analytics functions were commented as "non-blocking" but were
actually blocking the main execution flow by using .await, which waited
for HTTP requests to complete before continuing.

Changes:
- Update analytics function signatures to take owned values instead of references
- Use tokio::spawn at all call sites for fire-and-forget analytics
- Apply consistent pattern across `prover.rs`, `workers/online.rs`, and `workers/offline.rs`

Benefits:
- Proof verification and submission now happen at full speed
- Analytics HTTP requests run in background without delaying core functionality
- Failed analytics requests no longer impact prover reliability
- Comments now match actual non-blocking implementation

This ensures analytics are truly fire-and-forget as intended, improving
overall prover performance and user experience.
